perf(k3): the matmul landing streams at HBM rate
The CP8 128k prefill anatomy (deep rank 6,130 ms) charged 385 ms to
`k3_land`, the f32-partial -> bf16 landing after every dense projection.
The TileLang `land_batched` family landed one element per thread with
2-byte stores and reached 2.3 TB/s on the 12288-wide chunk landings
(550 us per launch at bucket 16896) — a third of HBM rate.

Replace the family with a hand-written CUDA kernel (`csrc/k3/k3_land.cu`):
eight columns per thread, two float4 loads per segment and one 16-byte
store, grid-stride over rows x n/8, batch a runtime value (no per-bucket
instantiation; rows not 16-byte aligned fall back to a scalar path). The
arithmetic is the retired kernel's spelling — ascending-segment f32 sum
from segment 0, one round-to-nearest-even cast — so at split_k = 1 the
landing is the bare cast and the output is bit-identical. The launcher
keeps its signature; no executor change.

Same-shape A/B (nsys, CP4 64k prefill gate, bucket 16896):
  nt=12288  n=12288  550 us -> 190 us  (1.25 GB, 6.6 TB/s)
  nt=18432  n=18432          287 us  (1.87 GB, 6.5 TB/s)
  nt=67584  n=33792          532 us  (3.4 GB, 6.4 TB/s)

A bf16 GEMM epilogue (cublasGemmEx with a bf16 C) was tried first and
rejected: cuBLAS picks a different kernel for a bf16 destination, so the
accumulation order moves and the 4-layer golden fixture flips at its
3-ULP-margin steps (same deviation distribution against the reference —
median 1 / p90 3 / max 12 bf16 ULP on both — but the gate's exact-match
contract is per kernel selection). The retired TileLang instantiations
(15 spans x 15 buckets) come off the build.

Gates (tray03, pruned 224-expert checkpoint): golden_decode 13/13 with
every step exact, paged_kv 3/3, spec_verify 6/6, cp_prefill 2/2.
