+++ pegainfer-kernels/KERNELS.md
@@ -131,7 +131,7 @@ f32, and narrowing any of them to bf16 measurably flips routing decisions.
 | op_id | Runtime owner | Rust wrapper | FFI symbol | Source (OUT_DIR `.cu`) | Backend | Shape / layout notes |
 | --- | --- | --- | --- | --- | --- | --- |
 | `k3.norm.rms_rbs` | `pegainfer-k3` | `ops::k3_rms_norm_rbs_batched_launch` | `k3_rms_norm_rbs_batched` | `k3_rms_norm_rbs_batched.cu` | TileLang AOT CUDA | KimiRMSNorm, round-before-scale: the normalized value lands bf16 *before* multiplying gamma. One row per block; `X`/`O` are `[b, h]` bf16 and gamma `[h]` is shared. h ∈ {7168, 512, 3584} (layer norms, MLA kv latent, routed latent) × 15 buckets; eps compiled in. |
-| `k3.linear.land` | `pegainfer-k3` | `ops::k3_land_batched_launch` | `k3_land_batched` | `k3_land_batched.cu` | TileLang AOT CUDA | Merge the column span `[off, off+n)` of each row's `[split_k, nt]` f32 partial and land bf16 once — the landing of every matmul. 15 (nt, n, off) spans × 15 buckets (14 from the sequential engine verbatim, plus `(12288, 12288, 0)` for chunked prefill's conv-input landing — the sequential engine casts conv inputs inside the conv kernel, the chunk lands them before its window build), split_k = 1 only — the single partial a framework GEMM (cuBLASLt, DeepGEMM) produces, where the merge degenerates to the slice and the cast. Masked tail store, so `n` need not divide 256. |
+| `k3.linear.land` | `pegainfer-k3` | `ops::k3_land_batched_launch` | `k3_land_cuda` | `csrc/k3/k3_land.cu` | CUDA | Merge the column span `[off, off+n)` of each row's `[split_k, nt]` f32 partial and land bf16 once — the landing of every matmul. Replaces the TileLang `land_batched` family (one element per thread, 2.3 TB/s on the chunked-prefill landings): eight columns per thread with 16-byte loads and stores, the same ascending-segment f32 sum and single round-to-nearest-even cast, so bit-identical. Shapes are runtime values (no per-bucket instantiation); rows not 16-byte aligned fall back to a scalar path. split_k = 1 at every launch site — the single partial a framework GEMM (cuBLASLt, DeepGEMM) produces. |
 | `k3.linear.land_rms_rbs` | `pegainfer-k3` | `ops::k3_land_rms_norm_rbs_batched_launch` | `k3_land_rms_norm_rbs_batched` | `k3_land_rms_norm_rbs_batched.cu` | TileLang AOT CUDA | `k3_land_batched` fused with the round-before-scale norm. One span — MLA q_a, `[0, 1536)` of the `14400`-wide fused projection — × 15 buckets, split_k = 1. |
 | `k3.elementwise.add2` | `pegainfer-k3` | `ops::k3_add2_batched_launch` | `k3_add2_batched` | `k3_add2_batched.cu` | TileLang AOT CUDA | `O = A + Bt` in bf16 addition (residual adds, routed + shared). n = 7168 × 15 buckets. |
 | `k3.elementwise.mul_sigmoid` | `pegainfer-k3` | `ops::k3_mul_sigmoid_batched_launch` | `k3_mul_sigmoid_batched` | `k3_mul_sigmoid_batched.cu` | TileLang AOT CUDA | `O = A * bf16(sigmoid(Bt))`, the MLA sigmoid output gate; the sigmoid is taken in f32 and lands bf16 before the product. n = 12288 × 15 buckets. |
