+++ pegainfer-kernels/src/ffi/k3_tilelang.rs
@@ -33,21 +33,7 @@ unsafe extern "C" {
         stream: CUstream,
     ) -> i32;
 
-    /// Merge the column span `[off, off + n)` of each row's `P [b, split_k, nt]`
-    /// f32 partial and land `O [b, n]` bf16 once. `split_k = 1` is the
-    /// single-partial case a framework GEMM produces.
-    pub fn k3_land_batched(
-        p: *const f32,
-        o: *mut c_void,
-        b: i32,
-        nt: i32,
-        n: i32,
-        off: i32,
-        split_k: i32,
-        stream: CUstream,
-    ) -> i32;
-
-    /// `k3_land_batched` followed by the round-before-scale norm against the
+    /// The matmul landing (`k3_land_cuda`) followed by the round-before-scale norm against the
     /// shared gamma `G [n]`.
     pub fn k3_land_rms_norm_rbs_batched(
         p: *const f32,
