+++ pegainfer-kernels/src/ops/k3/mod.rs
@@ -3,6 +3,7 @@
 mod deepgemm;
 mod flash_kda;
 mod flash_mla_prefill;
+mod land;
 mod mega_moe;
 mod mla_paged;
 mod moe_chain;
@@ -11,6 +12,7 @@ mod router_topk;
 pub use deepgemm::*;
 pub use flash_kda::*;
 pub use flash_mla_prefill::*;
+pub use land::*;
 pub use mega_moe::*;
 pub use mla_paged::*;
 pub use moe_chain::*;
